narrow public module facades
