+++ memory/src/lib.rs
@@ -9,8 +9,9 @@
 //! # Quick Start
 //!
 //! ```rust,ignore
-//! use swink_agent_memory::{JsonlSessionStore, SessionStore, SessionMeta};
-//! use swink_agent_memory::time::{now_utc, format_session_id};
+//! use swink_agent_memory::{
+//!     JsonlSessionStore, SessionMeta, SessionStore, format_session_id, now_utc,
+//! };
 //!
 //! let dir = JsonlSessionStore::default_dir().expect("config dir");
 //! let store = JsonlSessionStore::new(dir)?;
@@ -26,17 +27,17 @@
 //! store.save(&id, &meta, &messages)?;
 //! ```
 
-pub mod codec;
-pub mod compaction;
-pub mod entry;
-pub mod interrupt;
-pub mod jsonl;
-pub mod load_options;
-pub mod meta;
-pub mod migrate;
-pub mod store;
-pub mod store_async;
-pub mod time;
+mod codec;
+mod compaction;
+mod entry;
+mod interrupt;
+mod jsonl;
+mod load_options;
+mod meta;
+mod migrate;
+mod store;
+mod store_async;
+mod time;
 
 pub use compaction::{CompactionResult, SummarizingCompactor};
 pub use entry::SessionEntry;
