+++ plugins/web/src/content.rs
@@ -84,3 +84,66 @@ pub fn truncate_content(text: &str, max_len: usize) -> (String, bool) {
 pub fn is_html_content_type(content_type: &str) -> bool {
     content_type.contains("text/html") || content_type.contains("application/xhtml")
 }
+
+#[cfg(test)]
+mod tests {
+    use super::{extract_readable_content, is_html_content_type, truncate_content};
+
+    #[test]
+    fn truncate_content_short_text_no_truncation() {
+        let text = "Hello, world!";
+        let (result, truncated) = truncate_content(text, 100);
+        assert_eq!(result, text);
+        assert!(!truncated);
+    }
+
+    #[test]
+    fn truncate_content_long_text_is_truncated() {
+        let text = "a".repeat(1000);
+        let (result, truncated) = truncate_content(&text, 200);
+
+        assert!(truncated);
+        assert!(result.starts_with(&"a".repeat(160)));
+        assert!(result.ends_with(&"a".repeat(40)));
+        assert!(result.contains("[... content truncated"));
+    }
+
+    #[test]
+    fn html_content_type_detection_matches_expected_values() {
+        assert!(is_html_content_type("text/html"));
+        assert!(is_html_content_type("text/html; charset=utf-8"));
+        assert!(is_html_content_type("application/xhtml+xml"));
+        assert!(!is_html_content_type("application/json"));
+        assert!(!is_html_content_type("text/plain"));
+    }
+
+    #[test]
+    fn extract_readable_content_simple_article() {
+        let html = br#"
+        <!DOCTYPE html>
+        <html>
+        <head><title>Test Article</title></head>
+        <body>
+            <nav>Navigation links here</nav>
+            <article>
+                <h1>Main Heading</h1>
+                <p>This is the main article content that should be extracted by the readability algorithm. It contains enough text to be considered the primary content of the page.</p>
+                <p>Here is a second paragraph with more meaningful content to help the readability algorithm identify this as the main content block of the page.</p>
+            </article>
+            <footer>Footer content here</footer>
+        </body>
+        </html>
+        "#;
+
+        let url = url::Url::parse("https://example.com/article").unwrap();
+        let result = extract_readable_content(html, &url).unwrap();
+
+        assert_eq!(result.url, "https://example.com/article");
+        assert_eq!(result.content_type, "text/html");
+        assert_eq!(result.content_length, html.len());
+        assert!(!result.truncated);
+        assert_eq!(result.status_code, 200);
+        assert!(result.title.is_some());
+        assert!(result.text.contains("main article content"));
+    }
+}
