+++ plugins/web/src/lib.rs
@@ -5,22 +5,18 @@
 //! and extracting structured content. Includes safety policies for domain filtering,
 //! rate limiting, and content sanitization.
 
-pub mod config;
-pub mod content;
-pub mod domain;
-pub mod playwright;
-pub mod plugin;
-pub mod policy;
-pub mod search;
-pub mod tools;
+mod config;
+mod content;
+mod domain;
+mod playwright;
+mod plugin;
+mod policy;
+mod search;
+mod tools;
 
 pub use config::{SearchProviderKind, WebPluginConfig, WebPluginConfigBuilder};
-pub use content::{
-    ContentError, FetchedContent, extract_readable_content, is_html_content_type, truncate_content,
-};
+pub use content::FetchedContent;
 pub use domain::{DomainFilter, DomainFilterError};
-pub use playwright::{
-    ExtractedElement, ExtractionPreset, PlaywrightBridge, PlaywrightError, Viewport,
-};
+pub use playwright::{ExtractedElement, ExtractionPreset, Viewport};
 pub use plugin::{WebPlugin, WebPluginError};
 pub use search::{SearchError, SearchProvider, SearchResult};
