+++ plugins/web/src/policy/mod.rs
@@ -1,6 +1,6 @@
-pub mod domain_filter;
-pub mod rate_limiter;
-pub mod sanitizer;
+mod domain_filter;
+mod rate_limiter;
+mod sanitizer;
 
 pub use domain_filter::DomainFilterPolicy;
 pub use rate_limiter::RateLimitPolicy;
