+++ plugins/web/src/search/mod.rs
@@ -51,3 +51,18 @@ pub trait SearchProvider: Send + Sync {
         Box<dyn std::future::Future<Output = Result<Vec<SearchResult>, SearchError>> + Send + '_>,
     >;
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn search_result_is_publicly_constructible() {
+        let result = SearchResult {
+            title: "Rust".to_string(),
+            url: "https://www.rust-lang.org".to_string(),
+            snippet: "Systems programming language".to_string(),
+        };
+        assert_eq!(result.title, "Rust");
+    }
+}
