--- plugins/web/tests/common/mod.rs
@@ -1,1 +1,0 @@
-// Shared test helpers for swink-agent-plugin-web tests.
